requirements.txt
discord.py
requests
pandas
matplotlib
google-generativeai
